refactor: employer role type field cleanup
- in the `types/index.ts`, remove the `Employer` interface
- under the `types/` folder, create a `models/` folder for user role types
- move open job types and signup form fields into `types/auth/employer/`
- in `BestCompanies` and `FeaturedJobs` components, replace old `eid` field with `id`
- update import paths in:
  - `is-ilani` route
  - `filters` slice
  - `bestCompaniesApi`, `featuredJobsApi`, `jobDetailApi`
  - `jobDetail` type file
  - `employers`, `job-postings/job-detail` API route
  - `ApplicationModal`, `JobItem`, `JobList` components
- in `EmployerForm` component, rename some form field names for reusability:
  - `selectCity` → `city`
  - `selectDistricts` → `district`
  - `selectTaxOfficiesCity` → `taxCity`
  - `selectTaxOffice` → `taxOffice`
- update `EmployerSignupSchema` with the new form field names
- add `type-check` script in `package.json` to control type errors
