+++ app/api/firebase/employers/[slug]/route.ts
@@ -1,7 +1,7 @@
 import { collection, query, where, getDocs } from "firebase/firestore";
 import { NextResponse } from "next/server";
 import { db } from "../../firebaseConfig";
-import { Employer } from "@/types";
+import { Employer } from "@/types/auth/models/employer";
 
 export async function GET(
   _request: Request,
