+++ lib/redux/services/jobDetail.ts
@@ -1,4 +1,4 @@
-import { EmployerOpenJobs } from "@/types";
+import { EmployerOpenJobs } from "@/types/auth/employer/open-jobs.types";
 import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
 
 export const jobDetailApi = createApi({
