+++ types/auth/employer/signup-form.types.ts
@@ -1,0 +1,15 @@
+//? employer sign up form fields ?//
+export interface EmployerSignupFormFields {
+  nameAndSurname: string;
+  phone: string;
+  email: string;
+  password: string;
+  companyName: string;
+  taxNumber: string;
+  checkboxFirst: boolean;
+  checkboxSecond: boolean;
+  city: string;
+  district: string;
+  taxCity: string;
+  taxOffice: string;
+}
